Dockerfile: install at /seashell
This makes for a more predictable path than trying to use the home
directory.

I'm also temporarily removing sshpass. This chmod command won't work
anyway because there is no user called opam in this container. And in
general, it would be great to keep our Gorgonzola/Areesh-specific hacks
out of the generic Dockerfile.