Build an executable jar
This configuration for sbt-assembly uses a standard but silly trick to
build a jar that doubles as an executable command. It prepends a shell
script (with a shebang) that invokes the jar and makes the jar
executable.

This could obviate the need for the `fuse` script we provide, but I'm
not deleting it now to avoid disrupting anybody.